SSCS-8082 polish
